Test Cases for Dashboard and Table Component
Moved away from `enzyme` based test-suite to `@testing-library/react`
based test suite. This allows us to test `async` functionality.

We use `fake-indexdb/auto` component to provide `in memory` indexDB like
API to work so we don't have to mock any of the `idb` calls.

For testing the `Table` component, we need to provide database `db` to
the component, this is done through a `beforeAll` which resolves the
promise from `setupIndexedDB`. This is then passed as `db` property to
the table component.

All the test cases are now `async` because - all the functionality is
`async` -> When indexedDb promise is resolved, when DB Store is cleared
-> and so on.

We are skipping - Dashboard tests that check for `props` for now.

All the main tests pass. Good enough a starting point - known issues -

We see warnings like `component state update should happen inside
act(..)`, but right now we do not have a way to resolve them. For now
going ahead with this.